requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
DBUtils==3.1.0
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector==2.2.9
mysql-connector-python==8.4.0
mysql-to-sqlite3==2.1.12
packaging==24.0
PyMySQL==1.1.1
python-slugify==8.0.4
pytimeparse2==1.7.1
simplejson==3.19.2
tabulate==0.9.0
text-unidecode==1.3
tqdm==4.66.4
typing_extensions==4.12.1
Werkzeug==3.0.3
